changed int to double

+++ Edge.java
@@ -1,7 +1,7 @@
 public class Edge {
     private Vertex source;
     private Vertex destination;
-    private int weight;
+    private double weight;
 
     public Edge(Vertex source, Vertex destination, int weight) {
         this.source = source;
@@ -17,7 +17,7 @@ public class Edge {
         return this.destination;
     }
 
-    public int getWeight() {
+    public double getWeight() {
         return this.weight;
     }
 }
